fix(gpsjam): make the h3-conversion abort guard reachable + reject invalid cells (ce-code-review #4987)
Review finding (P2): the corrupt-data abort in processHexes compared h3Failures
against ALL CSV rows (lines.length-1 ≈ 47.6k), but h3Failures only increments
for rows that pass the minAircraft + interference filters and reach cellToLatLng
(~1.8k). So even 100% failure of attempted conversions could never exceed the
23.8k threshold — the guard was dead and corrupt hexes seeded as fresh.

Two-part fix:
- Denominator: count h3Attempts (rows that actually reach conversion) and abort
  on h3Failures > h3Attempts * 0.5 (guarded on h3Attempts>0).
- Validation: h3-js cellToLatLng silently returns a BOGUS centroid for non-hex
  garbage (only throws on hex-parseable-but-invalid cells), so validate isValidCell
  first — else garbage rows seed as fake hexes at a fixed location AND never count
  as failures. Now they're dropped and counted.

Extracted the parser to scripts/_gpsjam-parse.mjs (pure, minAircraft as a param)
so it's unit-testable without the seeder's main()/network path — the reviewer's
requested parser regression. Tests: normal parse + level classification, filter
drops, the attempted-denominator abort (3/3 fail throws despite 50 non-attempting
rows), and the minority-failure no-abort case. Fetcher verified end-to-end against
live gpsjam.org (934 high + 870 medium, unchanged).
